moved accounts.permissions to 'api' app
